+++ authors/apps/ah_notifications/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+
+from authors.apps.ah_notifications.views import NotificationAPIView
+
+app_name = "ah_notifications"
+
+urlpatterns = [
+    path('notifications/', NotificationAPIView.as_view(), name="ah_notifications")
+]
