+++ authors/apps/articles/signals.py
@@ -26,7 +26,6 @@ def send_notifications_to_followers(sender, instance, created, *args, **kwargs):
 def send_notifications_when_commented(sender, instance, created, *args, **kwargs):
 
     if instance and created:
-        # followers = [p.user for p in instance.author.followed_by.all()]
         users = [u.user for u in instance.article.users_favorites.all()]
         notify.send(instance, recipient=users,
                     verb='A new article have been published')
