+++ authors/apps/profiles/serializers.py
@@ -1,7 +1,4 @@
 from rest_framework import serializers
-from notifications.models import Notification
-from authors.apps.articles.models import Article
-from authors.apps.articles.serializers import ArticleSerializer
 
 from .models import Profile
 
@@ -22,19 +19,3 @@ class ProfileSerializer(serializers.ModelSerializer):
 
         return 'https://static.productionready.io/images/smiley-cyrus.jpg'
 
-
-class GenericNotificationRelatedField(serializers.RelatedField):
-    def to_representation(self, value):
-        if isinstance(value, Article):
-            serializer = (ArticleSerializer)
-
-        return serializer.data
-
-
-class NotificationSerializer(serializers.HyperlinkedModelSerializer):
-    target = GenericNotificationRelatedField(read_only=True)
-
-    class Meta:
-        model = Notification
-
-        fields = ('actor', 'target')
