+++ authors/apps/profiles/views.py
@@ -5,11 +5,10 @@ from rest_framework import serializers, status
 from rest_framework.views import APIView
 from rest_framework.exceptions import NotFound
 from notifications.models import Notification
-from rest_framework.renderers import JSONRenderer
 
 from .models import Profile
 from .renderers import ProfileJSONRenderer
-from .serializers import ProfileSerializer, NotificationSerializer
+from .serializers import ProfileSerializer
 from .exceptions import ProfileDoesNotExist
 
 import json
@@ -108,15 +107,3 @@ class FollowingAPIView(APIView):
         return Response(serializer.data, status=status.HTTP_200_OK)
 
 
-class NotificationAPIView(APIView):
-    permission_classes = (IsAuthenticated,)
-    serializer_class = NotificationSerializer
-    renderer_classes = (JSONRenderer,)
-
-    def get(self, request, username):
-        notifications = Notification.objects.filter(
-            recipient=request.user).all()
-        serializer = self.serializer_class(
-            notifications, many=True, context={'request': request})
-
-        return Response(serializer.data)
