+++ authors/urls.py
@@ -22,6 +22,8 @@ urlpatterns = [
 
     url(r'^api/', include('authors.apps.authentication.urls',
                           namespace='authentication')),
+    url(r'^api/', include('authors.apps.ah_notifications.urls',
+                          namespace='ah_notifications')),
     url(r'^api/profiles/', include('authors.apps.profiles.urls',
                                    namespace='profiles')),
